+++ src/neuroconv/datainterfaces/ecephys/baserecordingextractorinterface.py
@@ -225,6 +225,32 @@ class BaseRecordingExtractorInterface(BaseExtractorInterface):
             ]
             self.set_aligned_segment_timestamps(aligned_segment_timestamps=aligned_segment_timestamps)
 
+    def set_probe(self, probe, *, group_mode: Union[Literal["by_shank"], Literal["by_probe"]]):
+        """
+        Set the probe information via a ProbeInterface object.
+
+        Parameters
+        ----------
+        probe : probeinterface.Probe
+            The probe object.
+        group_mode : {'by_shank', 'by_probe'}
+            How to group the channels. If 'by_shank', channels are grouped by the shank_id column.
+            If 'by_probe', channels are grouped by the probe_id column.
+            This is a required parameter to avoid the pitfall of using the wrong mode.
+        """
+        # Set the probe to the recording extractor
+        self.recording_extractor.set_probe(
+            probe,
+            in_place=True,
+            group_mode=group_mode,
+        )
+        # Spike interface sets the "group" property
+        # Buy neuroconv expects "group_name" to be set
+        self.recording_extractor.set_property(
+            "group_name",
+            self.recording_extractor.get_property("group").astype(str)
+        )
+
     def align_by_interpolation(
         self,
         unaligned_timestamps: np.ndarray,